[PHPStanRules] Merge ForbiddenAssignInIfRule to ForbiddenMethodOrStaticCallInForeachRule
