+++ packages/phpstan-rules/tests/Rules/ForbiddenMethodOrStaticCallInForeachRule/Fixture/AssignInsideIf.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Symplify\PHPStanRules\Tests\Rules\ForbiddenAssignInIfRule\Fixture;
+namespace Symplify\PHPStanRules\Tests\Rules\ForbiddenMethodOrStaticCallInForeachRule\Fixture;
 
 function data()
 {
